+++ catalog/app/components/JsonEditor/ErrorHelper.js
