+++ catalog/app/containers/Bucket/PackageDialog/PackageDialog.js
@@ -9,7 +9,7 @@ import { parseJSON, stringifyJSON } from 'components/JsonEditor/State'
 import * as Notifications from 'containers/Notifications'
 import { useData } from 'utils/Data'
 import Delay from 'utils/Delay'
-import ErrorHelper from 'components/JsonEditor/Errors'
+import ErrorHelper from 'components/JsonEditor/ErrorHelper'
 import AsyncResult from 'utils/AsyncResult'
 import * as APIConnector from 'utils/APIConnector'
 import * as AWS from 'utils/AWS'
